+++ plugins/ui-react/src/components/index.ts
@@ -12,6 +12,7 @@ export * from './JsonHighlight';
 export * from './MultiplePicker';
 export * from './MultipleSelect';
 export * from './NotAvailable';
+export * from './PageHeaderActions';
 export * from './SingleSelect';
 export * from './StackedBarChart';
 export * from './StructuredMetadataList';